:rotating_light: Bring in widget to chart to keep knip happy
