+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/durationCategorical.tsx
@@ -1,5 +1,5 @@
 import {DurationUnit} from 'sentry/utils/discover/fields';
-import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/types';
+import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/common/types';
 
 export const sampleDurationCategoricalData: CategoricalSeries = {
   valueAxis: 'p99(transaction.duration)',
