+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/manyCategoriesCategorical.tsx
@@ -1,4 +1,4 @@
-import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/types';
+import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/common/types';
 
 export const sampleManyCategoriesData: CategoricalSeries = {
   valueAxis: 'count()',
