+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/plottables/categoricalDataSeries.tsx
@@ -8,7 +8,7 @@ import type {
   CategoricalItem,
   CategoricalSeries,
   CategoricalValueType,
-} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/types';
+} from 'sentry/views/dashboards/widgets/common/types';
 
 export type CategoricalDataSeriesConfig = {
   /**
