+++ static/app/views/dashboards/widgets/common/types.tsx
@@ -1,6 +1,7 @@
 import type {Confidence} from 'sentry/types/organization';
 import type {DataUnit} from 'sentry/utils/discover/fields';
 import type {ThresholdsConfig} from 'sentry/views/dashboards/widgetBuilder/buildSteps/thresholdsStep/thresholds';
+import type {PLOTTABLE_TIME_SERIES_VALUE_TYPES} from 'sentry/views/dashboards/widgets/common/settings';
 
 type AttributeValueType =
   | 'number'
@@ -83,6 +84,7 @@ type GroupBy = {
 
 // Aliases - allows divergence later if unique cases arise
 export type TimeSeriesGroupBy = GroupBy;
+export type CategoricalGroupBy = GroupBy;
 
 /**
  * Time series data. Unlike other time series abstractions, this is tightly supported by both the backend and the frontend. The `/events-timeseries/` endpoint uses this as the respone data, and `TimeSeriesWidgetVisualization` plottable objects accept this as the backing data.
@@ -143,3 +145,61 @@ export type Release = {
 };
 
 export type LegendSelection = Record<string, boolean>;
+
+/**
+ * The type of values in a categorical series.
+ * This is the same set of types supported by time series, but defined locally
+ * to avoid coupling to time series widgets.
+ */
+export type CategoricalValueType = (typeof PLOTTABLE_TIME_SERIES_VALUE_TYPES)[number];
+
+/**
+ * A single item in a categorical bar chart series.
+ */
+export interface CategoricalItem {
+  /**
+   * The category value for this data point.
+   */
+  category: string;
+  /**
+   * The numeric value for this category.
+   */
+  value: number | null;
+}
+
+/**
+ * Metadata for a categorical series.
+ */
+export interface CategoricalSeriesMeta {
+  /**
+   * The type of the values (e.g., "duration", "number")
+   */
+  valueType: CategoricalValueType;
+  /**
+   * The unit of the values, if applicable.
+   */
+  valueUnit: DataUnit | null;
+}
+
+/**
+ * A categorical data series for bar charts. Unlike time series,
+ * categorical series have discrete labels on the X axis rather than timestamps.
+ */
+export interface CategoricalSeries {
+  /**
+   * Metadata about the series.
+   */
+  meta: CategoricalSeriesMeta;
+  /**
+   * The aggregate function this series represents (e.g., "p95(span.duration)").
+   */
+  valueAxis: string;
+  /**
+   * The data points in this series.
+   */
+  values: CategoricalItem[];
+  /**
+   * Represents the grouping information for the series, if applicable.
+   */
+  groupBy?: CategoricalGroupBy[] | null;
+}
