+++ tests/js/fixtures/categoricalSeries.ts
@@ -1,4 +1,4 @@
-import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/types';
+import type {CategoricalSeries} from 'sentry/views/dashboards/widgets/common/types';
 
 export function CategoricalSeriesFixture(
   params: Partial<CategoricalSeries> = {}
